+++ flux/pom.xml
@@ -17,7 +17,7 @@
 
 
     <properties>
-        <java.version>1.15</java.version>
+        <java.version>1.8</java.version>
     </properties>
 
     <dependencies>
@@ -48,8 +48,8 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <configuration>
-                    <source>14</source>
-                    <target>14</target>
+                    <source>8</source>
+                    <target>8</target>
                 </configuration>
             </plugin>
         </plugins>
