+++ flux/src/main/java/com/lv/study/flux/lambda/StreamDemo.java
@@ -1,7 +1,6 @@
-package com.lv.study.lambda;
+package com.lv.study.flux.lambda;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
